Refactor: Move component actions to thunks and update related imports
- Deleted the actions.ts file and moved createComponent and loadComponents to thunks.
- Updated imports in populate-store.ts, store.ts, and test files to reflect the new thunks structure.
- Refactored hooks and selectors to improve loading state management for components.
- Cleaned up unused test cases in components-tab.test.tsx and create-component-form.test.tsx.
