+++ packages/packages/core/editor-components/src/populate-store.ts
@@ -1,7 +1,7 @@
 import { useEffect } from 'react';
 import { __dispatch as dispatch } from '@elementor/store';
 
-import { loadComponents } from './actions';
+import { loadComponents } from './thunks';
 
 export function PopulateStore() {
 	useEffect( () => {
